[#07][Refactoring] - modified delete event methods to also check for the last study/exam event and update course list accordingly - event update method now checks for the last study/exam and updates course list accordingly - as of now, once all course, study, and exam events for a specific course is deleted, this course will be removed from the student's course list
